Fix warning

+++ gtp.c
@@ -4978,18 +4978,18 @@ gtp_handler(struct gtp_trace_s *gts)
 		} else {
 #ifdef CONFIG_X86_32
 #if (LINUX_VERSION_CODE >= KERNEL_VERSION(2,6,25))
-			gts->x86_32_sp = &gts->regs->sp;
+			gts->x86_32_sp = (unsigned long)&gts->regs->sp;
 #else
-			gts->x86_32_sp = &gts->regs->esp;
+			gts->x86_32_sp = (unsigned long)&gts->regs->esp;
 #endif
 #endif
 		}
 	} else {
 #ifdef CONFIG_X86_32
 #if (LINUX_VERSION_CODE >= KERNEL_VERSION(2,6,25))
-		gts->x86_32_sp = &gts->regs->sp;
+		gts->x86_32_sp = (unsigned long)&gts->regs->sp;
 #else
-		gts->x86_32_sp = &gts->regs->esp;
+		gts->x86_32_sp = (unsigned long)&gts->regs->esp;
 #endif
 #endif
 	}
